♻️  (Jitsi Frame) allow configuration of the jitsi frame
Rename the Jitsi Component to Jitsi Frame to be more explicit.

Allow the configuration of the Jitsi Frame to apply different options
and access Jitsi api through the configure prop.
This should simplify the creation of different interfaces.
