+++ box-ui/src/components/tests/JitsiComponent.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import JitsiComponent from '../ts/JitsiComponent';
+import JitsiFrame from '../ts/JitsiFrame';
 
 test('renders learn react link', () => {
-    render(<JitsiComponent information={{ domain: '', roomName: '' }} />);
+    render(<JitsiFrame information={{ domain: '', roomName: '' }} options={{}} />);
     const linkElement = screen.getByText(/learn react/i);
     expect(linkElement).toBeInTheDocument();
 });
